+++ weblate/lang/tests.py
@@ -8,6 +8,7 @@ import warnings
 from gettext import c2py
 from io import StringIO
 from itertools import chain
+from unittest.mock import patch
 
 from django.core.management import call_command
 from django.test import TestCase
@@ -16,13 +17,14 @@ from django.urls import reverse
 from django.utils.translation import activate
 from weblate_language_data.aliases import ALIASES
 from weblate_language_data.languages import LANGUAGES
-from weblate_language_data.plurals import CLDRPLURALS, EXTRAPLURALS
+from weblate_language_data.plurals import CLDRPLURALS, EXTRAPLURALS, QTPLURALS
+from weblate_language_data.population import POPULATION
 
 from weblate.lang import data
 from weblate.lang.models import Language, Plural, PluralMapper, get_plural_type
 from weblate.trans.models import Unit
 from weblate.trans.tests.test_models import BaseTestCase
-from weblate.trans.tests.test_views import FixtureTestCase
+from weblate.trans.tests.test_views import FixtureTestCase, ViewTestCase
 from weblate.trans.util import join_plural
 from weblate.utils.db import using_postgresql
 
@@ -784,3 +786,104 @@ class PluralMapperTestCase(FixtureTestCase):
             )
         with override_settings(SIMPLIFY_LANGUAGES=False):
             self.assertEqual(arabic.get_aliases_names(), ["ar_ar", "ara", "arb"])
+
+
+class LanguageAliasesChangeTest(ViewTestCase):
+    """Simulate language aliases change in weblate_language_data and test results."""
+
+    old_code = "it"
+    new_code = "it_XX"
+
+    def setUp(self):
+        """Set up test environment."""
+        super().setUp()
+
+        def update_codes_in_dict(data: dict) -> dict:
+            """Replace old_code key with new_code in a language dict."""
+            copy = data.copy()
+            value = copy.pop(self.old_code)
+            copy[self.new_code] = value
+            return copy
+
+        def update_code_in_tuple(data: tuple) -> tuple:
+            """Replace old_code with new_code in a language tuple."""
+            new_data = []
+            for item in data:
+                if item[0] == self.old_code:
+                    item = (self.new_code, *item[1:])
+                new_data.append(item)
+            return tuple(new_data)
+
+        updated_aliases = ALIASES | {self.old_code: self.new_code}
+
+        patch_values = [
+            (
+                "weblate_language_data.languages.LANGUAGES",
+                update_code_in_tuple(LANGUAGES),
+            ),
+            (
+                "weblate_language_data.population.POPULATION",
+                update_codes_in_dict(POPULATION),
+            ),
+            ("weblate.lang.models.EXTRAPLURALS", update_code_in_tuple(EXTRAPLURALS)),
+            ("weblate.lang.models.QTPLURALS", update_code_in_tuple(QTPLURALS)),
+            ("weblate.lang.models.CLDRPLURALS", update_code_in_tuple(CLDRPLURALS)),
+            ("weblate_language_data.aliases.ALIASES", updated_aliases),
+            ("weblate.lang.data.ALIASES", updated_aliases),
+            ("weblate.lang.models.ALIASES", updated_aliases),
+        ]
+
+        for target, new_value in patch_values:
+            patcher = patch(target, new_value)
+            patcher.start()
+
+    def tearDown(self):
+        """Tear down after tests."""
+        super().tearDown()
+        patch.stopall()
+
+    def do_alias_language_update_and_check(
+        self, new_language_created: bool = True, old_language_deleted: bool = True
+    ):
+        """Test alias language update."""
+
+        def component_language_codes():
+            """Get current language codes for component translations."""
+            return [
+                translation.language.code
+                for translation in self.component.translation_set.all()
+            ]
+
+        self.assertIn(self.old_code, component_language_codes())
+        logs: list[str] = []
+        Language.objects.setup(True, logs.append)
+        if new_language_created:
+            self.assertIn(f"Created language {self.new_code}", logs)
+        else:
+            self.assertNotIn(f"Created language {self.new_code}", logs)
+
+        self.assertIn(self.new_code, component_language_codes())
+        if old_language_deleted:
+            # alias language should be deleted if blank
+            self.assertFalse(Language.objects.filter(code=self.old_code).exists())
+        else:
+            self.assertIn(self.old_code, component_language_codes())
+            self.assertTrue(Language.objects.filter(code=self.old_code).exists())
+
+    def test_update_language_alias(self):
+        """Test simple alias change."""
+        self.do_alias_language_update_and_check()
+
+    def test_update_language_alias_no_deletion(self):
+        """Test alias change with no deletion of old language."""
+        self.component.new_lang = "add"
+        self.component.new_base = "po/hello.pot"
+        self.component.save()
+        it_xx = Language.objects.create(code="it_XX", name="New Italian")
+        it_xx.plural_set.create(
+            number=0,
+            formula="0",
+            source=Plural.SOURCE_DEFAULT,
+        )
+        self.component.add_new_language(it_xx, None)
+        self.do_alias_language_update_and_check(False, False)